Corrected a minor mistake
Previously, I set the data in our dummy node using the data passed to
the removeHelper() method. The technically correct method of setting
this data is to use the data from the current node. At this point in the
code, the values are the same but the data objects are not.