Fixed a bug in the remove method
Previously, I did not account for balancing operations we must perform
when we remove a node with two children and replace it with its
successor. Namely, we must balance the node that was the parent of the
successor as well as the successor after we have placed it in its new
location in the tree.